Remove process based installer and renamed service based installer to `WindowsNeo4jInstaller`
